requirements.txt
boto3==1.12.29
numpy==1.18.2
opencv-python-headless==4.2.0.32
streamlink==1.3.1
pytz==2019.3
redis==3.4.1
tensorflow==1.14.0
Keras==2.2.4
Pillow==5.3.0